requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
aiohttp==3.9.5
aiogram==3.2.0
numpy==1.26.4
torch==2.3.1
cryptography==42.0.5
pydantic==2.5.3
pillow==10.3.0
moviepy==1.0.3
